Refactored tests and added another scheduler test
