+++ herc/aurorastub.py
@@ -3,8 +3,8 @@
 class AuroraStub(object):
     """Mock Aurora backend that returns correct-looking data."""
 
-    def __init__(self):
-        pass
+    def __init__(self, fakestatuspath = "data/stub_aurora_jobstatus.json"):
+        self.fakestatuspath = fakestatuspath
 
     def requestjob(self, jobid, jobrq):
         """No return value required; nothing to do."""
@@ -12,5 +12,5 @@ class AuroraStub(object):
 
     def status(self, jobid):
         """Returns a good-looking placeholder job status."""
-        with open("data/stub_aurora_jobstatus.json", 'r') as fakestatus:
+        with open(self.fakestatuspath, 'r') as fakestatus:
             return fakestatus.read()
